Updated the Stackable keys to have them start with a root slash.
This was brought up on the mailing list and seemed like a reasonable
idea. It also will make it easier for implementing libraries to use one
class for systems that use stackable and non-stackable components,
making it easier to segment the two pools (internal namespacing or
something) as there will *always* be a slash when Stackables are used.