Remove vendor/ from PSR-0 examples
This has come up in conversation a few times, where some users were making a direct connection between Composer and PSR-0 due to having the "vendor/" directory in there.

This to me is a trivial change, which might help reduce some confusion. They think that EVERYTHING inside the vendor/ folder should be PSR-0, which if you're using Composer it wont be, as Composer <vendor-name>/<package-name> is not PSR-0 compatible.

I'll throw a vote on the mailing list to try and get this one approved, as my last PR earned me a dressing down.