Removed duplicate interfaces
There were example interfaces that matched those in the proposal but
were separate to help with distribution. However, with the continued
changes being made the interface files outside of the actual proposal
have been been kept up to date.

Once the proposal is actuall finalized the interfaces will need to be
distributed, but that is most likely going to be under a different
repository.