Do not refer to Fully Qualified Class Name in the PSR
Last week, I [created and shared a gist](https://gist.github.com/AmyStephen/6943532) that explains why there is no reason for the PSR to ever talk about a FQCN. Using the gist, you can verify that PHP will not ever pass the FQCN into an spl_autoloader. A FQCN is internal to PHP which passes the QCN into the spl_autoloader. 

In my next commit, I will adapt the language of the PSR to use QCN when I add that definition.