Do not redefine the term Scheme which is already associated with URI scheme and RFC3986
Scheme is a well defined and heavily used term closely associated with a [URI](http://tools.ietf.org/html/rfc3986). It should not be used to describe what is essential differences between Windows and other operating systems related to a directory separator. 

I will remove the word scheme in the body of the PSR in a later commit. It really is not necessary to define what the autoloader does. If we focus on input and output then it really won't be necessary to discuss this process.