Simplify resource path term
Again, the terms are not describing process but rather a point of reference for later when the mapping is described. Like the example, left that and added one to resource path. IMO, should not include the mapping process -- it's only a definition of resource path - not a mapping process. Keep it simple. Nice and easy.