Yadda, yadda, yadda
Please. If you are into a preamble, we are in the deep grass. Take a mulligan. We want people to read this sucker. (I think the preamble was needed because too many ideas were brought into a tight spec and it no longer made sense.) If we strip it back down to it's skivvies, we can avoid a preamble.