Terse
Added official PHP term *unqualified class name* and DRYed out the specification. If it's in the terms, it need not be repeated elsewhere. If it's in the specification, it need not be repeated in the terms, and so forth.

Also, recommend all three parts of the examples (which are great), be moved to a separate location where PR's can be made over time and within the rules of FIG.

These changes strip the language way down without changing the rules. I realize a lot of combining took place to bring everyone along but it ended up dulling the meaning IMO. 

I split the commits to help illustrate points. Combined these represent significant change and I understand the challenge the editors are facing. Take this for what it's worth and thank you for your consideration of these ideas.