Reverted @type to @var and changed compound @var notation to match PHPs
During discussions is became obvious to me that the proposed rename of
@var to @type encountered quite a bit of resistance and posed difficulties
with existing annotation consumers, such as Doctrine, because they are
using that namepart in their annotation names.

In addition, due to a discussion at http://youtrack.jetbrains.com/issue/WI-7659,
it become obvious that the notation where a DocBlock preceeds an individual
property in a compound statement that it is picked up by PHP's Reflection
library. As such I adapted the spec to match PHP's behaviour.

You can reproduce this using the following script:

    <?php

    class A
    {
        public
            /** This is a DocBlock */
            $b,
            /** This is a second DocBlock */
            $c;
    }

    $reflectionProperty = new ReflectionProperty('A', 'b');
    var_dump($reflectionProperty->getDocComment());
    $reflectionProperty = new ReflectionProperty('A', 'c');
    var_dump($reflectionProperty->getDocComment());

When you execute this you get the following output:

    string(25) "/** This is a DocBlock */"
    string(32) "/** This is a second DocBlock */"