[PSR-7] Revert to original(-ish) interfaces
This patch removes the client/server interface segregation, and re-instates full
mutability (with two notable exceptions). The rationale was discussed on the
mailing list; essentially, in practice, mutability will be simpler for consumers
and allow more varied architectures.

The interfaces presented now are:

- `StreamableInterface`
- `MessageInterface`
- `RequestInterface`
- `ServerRequestInterface` (was `IncomingRequestInterface`)
- `ResponseInterface`

The request interface also now has a "base URL" member. The "URL" member is only
used to store the path and query string, while the base URL is intended for the
scheme, host, port, and authentication information.

Fixes #296
Fixes #394