[#453] Use `#`-prefixed headers instead of separator-style
Per #453, using separator-style headers (those with `===` and `---` filled lines
following the header) can cause rendering issues if the header starts with a
`{number}.` prefix. All other proposals use `#`-prefixed headers, likely to
address this very issue, so this patch alters headers to do the same both for
consistency and to correct rendering issues.