[PSR-7] Errata: scheme delimiters
A [discussion on the mailing list](https://groups.google.com/d/msgid/php-fig/db3d5421-30a5-4135-9f9e-fe432f03d719%40googlegroups.com?utm_medium=email&utm_source=footer) raised the point that
`:` is the actual scheme delimiter. In reviewing [RFC 3986 section 3](https://tools.ietf.org/html/rfc3986#section-3),
I noted that `:` is the end delimiter for a scheme, and `//` is the opening
delimiter for the authority. Since the `UriInterface` does not allow setting the
authority in a single go, the only time the `//` delimiter will likely be used
is as part of the scheme, making the current wording valid. However, the wording
should also note that a trailing `:` delimiter MUST be removed as well.

Created an "Errata" section that adds this; we can vote on it post-acceptance.