Alterative way to address consistency issues mentioned in #491
This PR suggests the following changes:

- `getHeader($name)` should returns the header value(s) as set by the user; it does not return a concatenated string.

- `getHeaders()` is now essentially the equivalent of calling `getHeader()` for every header in the message; it does not return an array of header line strings.

- `getHeaderLine($name)` is a new method that takes the place of the old `getHeader()`, where the return value is expected to be an entire header line with the values concatenated into a comma-separated string.

- `getHeaderLines()` now takes the place of the old `getHeaders()`; it is essentially the equivalent of calling `getHeaderLine()` for every header in the message.

This PR has no equivalent for the old `getHeaderLines($name)` behavior that would return an array of multiple lines for a single header, each with a separate value for that header. The user can mimic that behavior with something like:

    $lines = [];
    foreach ($message->getHeader($name) as $values) {
        foreach ((array) $values as $value) {
            $lines[] = "{$name}: $value";
        }
    }

Finally, this PR does not enforce the idea that `getHeader()` or `getHeaders()` should return an array for the header values themselves. This is to allow a further bit of consistency:

    $message = $message->withHeader('Foo', 'foo-value-1');
    assert($message->getHeaders()['Foo'] === $message->getHeader('Foo'));

    $message = $message->withAddedHeader('Foo');
    assert($message->getHeaders()['Foo'] === $message->getHeader('Foo'));

I think this is both more flexible and more consistent than the previous version, and addresses both my own gripes and those expressed in #491.