Added getStream() and move() to UploadedFileInterface
Per m00t, this patch adds `getStream()` and `move()` as preferred methods for
interacting with the uploaded file. `getStream()` will return a valid PHP stream
resource, while `move()` will move the file upload to the designated filesystem
path.

`getStream()` can be used for stream operations, such as
`stream_copy_to_stream()`, which can be useful for such purposes as copying the
file upload to a cloud storage solution. It can also be used for manually
copying the stream contents to a file if desired.

I chose NOT to return a StreamableInterface instance, as I think in this
particular case, PHP stream operations will be the norm, and returning a
ready-to-use PHP stream will make the most sense.