Common sense bylaw caveats on member expulsion
The FIG does not need to get into the habit of making bylaws for every
decision the secretaries end up needing to make, but we can certainly
keep the bylaws up to date with some of the precedence being set.

Sure, some people will worry about some authoritarian group trying to
take control - or something, but really this specific caveat should be
common sense. Let's not panic about slippery slopes, and quite simply
let's add this rule which should have been there since the start.

A bouncer doesn't get to ask you if you think you should be kicked out
of a bar, and an potentially expelled member doesn't need a vote on their
exclusion. They don't get to vote on their membership vote to be
included either... So long as they respect the self throttling rules, they'll have plenty of time to put their points forward during the discussion
vote. But really, this shouldn't happen often. Maybe once or twice.

Consider this a solution to the disucssion thread started
[here](https://groups.google.com/forum/#!topic/php-fig/gMnU02lU1D0).