Removed Developer Advocacy as a requirement for secretaries
I don't think that submitting talks about the FIG, blogging on it, or any of the other roles a developer advocate does should be a hard requirement for a secretary. I'm sure a secretary can do an amazing job without needing to be a cheerleader for the group, so I am suggesting we remove this requirement.

I will start the discussion thread on the mailing list.