Updates to EventManager component proposal
I don't know the progress of this EventManager (PSR-14) proposal, whether it is even under any consideration or not, but, I do feel that it is very important building block of applications and should be standardized.

However, the current state of the proposed API is very weak in my opinion to be considered as PSR standards.  I don't know where to have discussion about this, if it open for discussion (which I think should be at this stage).

1. `EventManagerInterface`
1.a) Lets start with naming, `EventManagerInterface` is not a good name for standards.  I don't think PSR should try to standardize any "manager".  Usually 'Manager' prefix is used throughout application to define various services.  However, if we change the name to something like `EventNotifierInterface` (or `EventDispatcherInterface`) this would be a better name for standards.  Which leads me to next suggestions:

1.b) I particular favor `EventNotifierInterface` because of the nature of the component.  It is to be used as 'notification center' as primary job, rather than 'managing' events (which is secondary).  

1.c) I also think `notify` is better naming for this instead of `trigger`.  Trigger has some kind of user action, or explicit demand (like triggering errors).  However, 'notify' sounds more softer and flows with the purpose of the component.

1.d) It also have changes requested in the PR #833, removing the clearListener() method.

2. `EventInterface`
2.a) Most of the suggestion for this interface goes along with my previous PR #833.  PSR should try to have minimum number of methods in its API definitions.  Therefore following methods are not necessary for the interface:
2.a.i `getParam(string)`
2.a.ii `setParams(array)`
2.a.iii `setTarget(mixed)`
2.a.iv `setName(string)`
There are many ways to create and set these values.  Interface should not need to define these.  One can choose not to have these methods such as `setName` and `setTarget` defined, rather use constructor have them assigned and mandate such.  As for convenient method like `getParam(string)`, this is also really not necessary.

2.b) The method `stopPropagation` should NOT take any argument.  It doesn't make much sense to have one listener stopping the propagation, and other event (or any where else) starting again, which would be counter to the intent of the method to begin with.  Any listener when decides to stop the event chain, it will simply call stopPropagation().

I have been spending a LOT of time lately with events related component by doing researching various frameworks, and platforms.  As with most things, there are MANY ways to accomplish this.  But I would like to see PSR or other standard community standardize this component.