`declare` must follow the rule for assignment operator
`declare` statements must follow the rule for assignment operator which dictates to have exactly 1 space around the operator.

There was no reason to make it an exception. It made the rule harder to explain, and harder to remember. By keeping a single rule for all cases everything is made simpler.