Refactores CORS example to instead address Authorization
As @shadowhand points out, CORS is a domain many specification readers
may not be familiar with. As such, refactored it to be an authorization
middleware, with the following behavior:

- If no authorization is required, it is a pass-through proxy to the
  handler.
- If authorization is required, and the request is invalid, it returns
  an "unauthorized" response.
- Otherwise, it passes the request to the handler, and signs the
  response returned.

This approach hits the same points I was trying to make with the CORS
example, but with a domain more people will be familiar with.