Proposed style rule for language constructs
where parentheses are optional. These are generally easy to identify in the documentation as they will have a note like:

>echo is not actually a function (it is a language construct), so you are not required to use parentheses with it. 

The real goal of this style rule is to get consistent usage of parentheses or sans-parentheses within a given codebase, but since consistency across all code greatly increases readability, we should probably choose a favorite, and I'd go with no parentheses as it looks a bit cleaner to me and is fairly consistent with how most developers use `echo`. It also emphasizes that the keyword is not a function.

This could also be seen as an extension of the PSR-2 rule on Control Structures :

> Control structure keywords MUST have one space after them; method and
function calls MUST NOT.

when using parentheses for language constructs, it should follow the style of a function. When not using parentheses, it should follow the style of a control structure keyword.