Proposed changes for uploaded file factory
First major change is that it MUST extend StreamFactoryInterface. This is
required, as uploaded files are expected to _compose streams_. As such,
they need a way to convert a file name or resource into a stream.

Second major change is allowing a _StreamInterface implementation_ as a
potential type for the `$file` argument. When that is present, we can
use that information to determine the uploaded file size, but it also
means we do not need to create a stream instance immediately.

Finally, removed the `fwrite($resource, $body);` line, as it does not
make sense in this context.