Update ResponseFactoryInterface to include $reasonPhrase
Per WG discussion, we are adding an optional `string $reasonPhrase`
argument. The rationale is that the only way to provide a reason phrase
to a response is via the `withStatus()` method, which requires both the
status code and reason phrase. As such, the factory should mimic that
signature, as a status code is the only required element of a new
response.

Additionally, this patch adds in PHP 7 type hints.