Simplify ServerRequestFactoryInterface
Removes the `createServerRequestFromArray` method, and adds the optional
argument `$serverParams` to the `createServerRequest` method. Since
server params cannot be injected via a mutator method, users MUST be
able to pass them to the factory. Additionally, these parameters COULD
omit values normally seen in mod_php, mod_cgi, and mod_fpm SAPIs that
would allow generating a (mostly) populated request instance. As such,
the WG decided that the method and URI should remain required, and that
the two signatures should be combined.

This patch also updates the signature to PHP 7 typehints.