Update numerical index to use statuses directly
The indirection of these one-letter cryptic abbreviations does not
seem to add value to the reader. The cost of having a ever so slightly
wider column and having to type a word rather than a character seems
to outweigh the benefit of readability and reduced confusion.

Also re-order the columns so that Status is last.