[PSR-17] Removes/rephrases "popular"
A number of people have indicated they are uncomfortable with the word
"popular" when describing various PSR-7 implementations in the meta
document; no metrics were provided, and popularity may change over the
lifetime of the spec.

This patch rephrases all occurrences of the word to more accurately
describe the various projects.