PSR-17 acceptance
As of 2018-07-29, the [last vote has been cast for PSR-17](https://groups.google.com/d/msg/php-fig/M8PapGXXE1E/WuQlZ3trBAAJ),
and the specification has unanimous approval from the Core Committee.

This patch updates the status of the proposal to "Accepted", and moves
the documents under the `/accepted/` tree of the repository.