Enforce `elseif` instead of  `else if`
This has a very low visual impact and is trivial to apply to existing code. So instead of leaving the specification opened to preferences, it is more beneficial to enforce our choice.

This also remove the only `SHOULD` that is not about line length, making the specification less about possibilities and more about enforced rules.