Completely revert tactic
From advocating the use of local times (which is hard to do for local
times) this argumentation moves towards explaining that the interface
itself, the explicit contract, does not care about the implementation
and with this rather open explicit contract everyone can also implement
their own implicit contract to use "UTC" or whatever they want. If they
want to. The interface itself is as open as possible while at the same
time being as strict as necessary