feat: document language-level parameter and return type declarations
This patch adds section 7.2 to the metadocument, detailing the (proposed) changes to the specification to use explicit, language-level parameter and return type declarations.
It documents the addition of parameter type declarations for the 1.1 release, the addition of return type declarations for 2.0, and how implementations may adopt one or both versions.
It also documents and links to the (forthcoming) package releases within the primary document, detailing the changes in brief.

The format follows that of the already accepted PSR-13 evolution proposal.